pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "setuptools-scm>=8.1"]
build-backend = "setuptools.build_meta"

[project]
name = "qr-sampler"
dynamic = ["version"]
description = "Plug any randomness source into LLM token sampling"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "qr-sampler contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
readme = "README.md"
dependencies = [
    "numpy>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.5.0",
    "grpcio>=1.68.0",
    "protobuf>=5.26.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/alchemystack/qr-sampler"
Repository = "https://github.com/alchemystack/qr-sampler"
Issues = "https://github.com/alchemystack/qr-sampler/issues"
Changelog = "https://github.com/alchemystack/qr-sampler/blob/main/CHANGELOG.md"

[project.scripts]
qr-sampler = "qr_sampler.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "scipy>=1.12.0",
    "ruff>=0.8.0",
    "mypy>=1.10.0",
    "pre-commit>=4.0",
    "bandit>=1.8.0",
    "types-PyYAML>=6.0",
    "click>=8.0",
    "jinja2>=3.1",
]
cli = [
    "click>=8.0",
    "jinja2>=3.1",
]
openentropy = [
    "openentropy>=0.10.0",
]
vllm = []
vllm-metal = []

[project.entry-points."vllm.logits_processors"]
qr_sampler = "qr_sampler.processor:QRSamplerLogitsProcessor"

[project.entry-points."qr_sampler.engine_adapters"]
vllm = "qr_sampler.engines.vllm:VLLMAdapter"

[project.entry-points."qr_sampler.entropy_sources"]
system = "qr_sampler.entropy.system:SystemEntropySource"
quantum_grpc = "qr_sampler.entropy.quantum:QuantumGrpcSource"
timing_noise = "qr_sampler.entropy.timing:TimingNoiseSource"
mock_uniform = "qr_sampler.entropy.mock:MockUniformSource"
openentropy = "qr_sampler.entropy.openentropy:OpenEntropySource"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"qr_sampler" = ["profiles/**/*.yaml", "templates/**/*.j2", "py.typed"]

[tool.setuptools_scm]
fallback_version = "0.1.0.dev0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",       # pyflakes + pycodestyle
    "I",                  # isort
    "N",                  # pep8-naming
    "UP",                 # pyupgrade
    "B",                  # flake8-bugbear
    "SIM",                # flake8-simplify
    "TCH",                # flake8-type-checking
    "RUF",                # ruff-specific rules
]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
packages = ["qr_sampler"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["grpc", "grpc.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["torch", "torch.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openentropy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["click", "click.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["jinja2", "jinja2.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/qr_sampler"]
omit = [
    "src/qr_sampler/proto/*",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
